create new component folder to organize calculator
